feat: create BeforeStartGame.tsx and move there components from sprinter
